feat(campushop): ajustes de abas, carrinho, pedidos, conta e chat

+++ Documents/CAMPUSHOP/src/main/java/br/com/campushop/campushop_backend/repository/UsuarioRepository.java
@@ -1,13 +1,14 @@
 package br.com.campushop.campushop_backend.repository;
 
-import br.com.campushop.campushop_backend.model.Usuario;
+import br.com.campushop.campushop_backend.model.Usuario; // Importando a classe Usuario para usar no Optional e no método de busca por email
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
+public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
 
     Optional<Usuario> findByEmail(String email);
